Move Guice module to a new package
